fix the sigma parameter in double (precedently in size_t)

+++ libraries/terry/src/terry/sampler/gaussian.hpp
@@ -16,7 +16,7 @@ namespace sampler {
 struct gaussian_sampler
 {
 	size_t size;
-	size_t sigma;
+	double sigma;
 	gaussian_sampler()
 	{
 		size  = 3.0;
